Final Commit
This is the final commit and what I did in the final commit was add a line of comment to the index.html. Then, I added in some css code for the input boxes and the input, select, and textarea to make them focused and valid/invalid and what not and after that code I finished my comments and completed the project which is now ready for review.